refactor(header): improved code modularity
