+++ frontend/app/components/NavBar/components/Logo.tsx
@@ -1,0 +1,17 @@
+import Image from "next/image";
+import Link from "next/link";
+
+export const Logo = () => {
+  return (
+    <Link href={"/"} className="flex items-center gap-4">
+      <Image
+        className="rounded-full"
+        src={"/logo.png"}
+        alt="Quivr Logo"
+        width={48}
+        height={48}
+      />
+      <h1 className="font-bold">Quivr</h1>
+    </Link>
+  );
+};
