+++ frontend/app/components/NavBar/components/NavItems/index.tsx
@@ -2,19 +2,24 @@
 import { useSupabase } from "@/app/supabase-provider";
 import { cn } from "@/lib/utils";
 import Link from "next/link";
-import { Dispatch, FC, HTMLAttributes, ReactNode, SetStateAction } from "react";
+import { Dispatch, FC, HTMLAttributes, SetStateAction } from "react";
 import { MdSettings } from "react-icons/md";
-import Button from "../ui/Button";
-import DarkModeToggle from "./DarkModeToggle";
+import Button from "../../../ui/Button";
+import { AuthButtons } from "./components/AuthButtons";
+import { DarkModeToggle } from "./components/DarkModeToggle";
+import { NavLink } from "./components/NavLink";
 
 interface NavItemsProps extends HTMLAttributes<HTMLUListElement> {
   setOpen?: Dispatch<SetStateAction<boolean>>;
 }
 
-const NavItems: FC<NavItemsProps> = ({ className, setOpen, ...props }) => {
+export const NavItems: FC<NavItemsProps> = ({
+  className,
+  setOpen,
+  ...props
+}) => {
   const { session } = useSupabase();
   const isUserLoggedIn = session?.user !== undefined;
-  const isLocal = process.env.NEXT_PUBLIC_ENV === "local";
 
   return (
     <ul
@@ -24,7 +29,7 @@ const NavItems: FC<NavItemsProps> = ({ className, setOpen, ...props }) => {
       )}
       {...props}
     >
-      {isLocal || isUserLoggedIn ? (
+      {isUserLoggedIn ? (
         <>
           <NavLink setOpen={setOpen} to="/upload">
             Upload
@@ -62,32 +67,10 @@ const NavItems: FC<NavItemsProps> = ({ className, setOpen, ...props }) => {
             </Link>
           </>
         )}
-        {!isLocal && !isUserLoggedIn && (
-          <Link href={"https://try-quivr.streamlit.app"}>
-            <Button variant={"secondary"}>Try Demo</Button>
-          </Link>
-        )}
+        {!isUserLoggedIn && <AuthButtons />}
+
         <DarkModeToggle />
       </div>
     </ul>
   );
 };
-
-interface NavLinkProps {
-  children: ReactNode;
-  to: string;
-  setOpen?: Dispatch<SetStateAction<boolean>>;
-}
-
-const NavLink: FC<NavLinkProps> = ({ children, to, setOpen }) => {
-  return (
-    <li className="group relative">
-      <Link onClick={() => setOpen && setOpen(false)} href={to}>
-        {children}
-      </Link>
-      <hr className="aboslute top-full border border-transparent border-b-primary dark:border-b-white scale-x-0 group-hover:scale-x-100 group-focus-within:scale-x-100 transition-transform" />
-    </li>
-  );
-};
-
-export default NavItems;
