+++ frontend/app/layout.tsx
@@ -4,7 +4,7 @@ import { Inter } from "next/font/google";
 import { cookies, headers } from "next/headers";
 import { BrainConfigProvider } from "../lib/context/BrainConfigProvider/brain-config-provider";
 import Footer from "./components/Footer";
-import NavBar from "./components/NavBar";
+import { NavBar } from "./components/NavBar";
 import { ToastProvider } from "./components/ui/Toast";
 import "./globals.css";
 import SupabaseProvider from "./supabase-provider";
